requirements.txt
py-cord
python-dotenv
nextpy